entities are now just ids as they should be - moved all meta data about an entity to a struct EntityMetaData that is saved separately. started integrating saving of scenes added scene and entity definition resources (wip) types can now be retrieved by strings (wip - the current solution isn't very great...)

+++ Dawn/Dawn/src/Dawn/UI/Editor/imgui_debug.cpp
@@ -10,7 +10,7 @@ namespace Dawn
 	{
 		auto Settings = g_Application->GetSettings();
 		auto Renderer = g_Application->GetRenderer();
-		static u32 yOffset = 20;
+		static float yOffset = 20.0f;
 
 		ImGui::Begin("GBuffer_Position", 0, ImGuiWindowFlags_NoTitleBar |
 			ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove |
